changed model variable to item

+++ app/src/controller/base-controller.ts
@@ -15,8 +15,8 @@ export default class BaseController {
 
     create = (req: Request, res: Response) => {
         this._model.create(req.body)
-            .then(model => {
-                return dataResponse(res, model);
+            .then(item => {
+                return dataResponse(res, item);
             })
             .catch(ValidationError, err => {
                 validationErrorResponse(res, err);
@@ -24,25 +24,30 @@ export default class BaseController {
     };
 
     update = (req: Request, res: Response) => {
-        this._model.findByPk(req.params.id).then(model =>
-            model.update(req.body)
-                .then(model => {
-                    return dataResponse(res, model);
-                })
-                .catch(ValidationError, err => {
-                    validationErrorResponse(res, err);
-                })
+        this._model.findByPk(req.params.id).then(item => {
+                // Todo: consider moving this to a middleware
+                if (item.userId !== req.user.id) {
+                    return errorResponse(res, 'This model belongs to another user.', statusCode.FORBIDDEN);
+                }
+                return item.update(req.body)
+                    .then(item => {
+                        return dataResponse(res, item);
+                    })
+                    .catch(ValidationError, err => {
+                        validationErrorResponse(res, err);
+                    })
+            }
         );
     };
 
     destroy = (req: Request, res: Response) => {
-        this._model.destroy({where: {uuid: req.params.id}})
+        this._model.destroy({where: {uuid: req.params.id, userId: req.user.id}})
             .then(deleted => {
                 if (deleted !== 1) {
                     return errorResponse(res, 'Not Found', statusCode.NOT_FOUND)
                 }
                 return messageResponse(res, 'Deleted');
-                })
+            })
             .catch(Error, err => {
                 errorResponse(res, err);
             });
